[Feature] Add padTime and sendNotification utilities; update SolCycle and ResetButton components for improved time handling and notifications
