+++ utils/padTime.ts
@@ -1,0 +1,5 @@
+const padTime = (time: number) => {
+  return time.toString().padStart(2, "0");
+};
+
+export default padTime;
